requirements.txt
# Core Dependencies
deepgram-sdk>=3.0.0
loguru>=0.7.0
numpy>=1.24.0
PyQt6>=6.6.0
sounddevice>=0.4.6
pynput>=1.7.6
pyautogui>=0.9.54
pyperclip>=1.8.2
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0

# Build
pyinstaller>=6.3.0
